Disable RGB Underglow on start
RGB seems to be a contributing factor in USB device crashes.

Disabling RGB with the hot key seems to be only partially working.
Sometimes the RGB turns off, sometimes only half turn off.

If the board has the RGB setting saved as off, the board boots with the
LEDs turned on, but the feature is disabled and does not animate -
entering a sort of half off state. Disabling RGB Underglow by default on
start fixes this. However, this half enabled state is significantly more
stable than having RGB actually turned on, perhaps this stability issue
is related to the ws2812 PIO driver, or a conflict with RGB underglow
tick causing USB to be delayed.