Add custom display code
This adds a minimal custom display system using the character
framebuffer library from Zephyr instead of ZMK's display system on top
of LVGL. This causes the USB stack to become unstable in a very similar
manner to enabling RGB, but proves that the display hardware is
functional. While testing, it seems the UART logging also significantly
contributes to instability.